requirements.txt
python-dotenv
requests
fastapi
pydantic
uvicorn